move test case
