--- src/range/__tests__/prediction.json
@@ -1,1 +1,0 @@
-[{"atomIDs":["8"],"diaIDs":["did@`@fTfYea`H`@GzP`HeT"],"integral":1,"delta":6.853,"atomLabel":"H","j":[{"assignment":"9","diaID":"did@`@f\\bbbQ[hBB@A~dHBIU@","coupling":1.564,"multiplicity":"d"},{"assignment":"10","diaID":"did@`@fTfUff`@h@GzP`HeT","coupling":0.519,"multiplicity":"d"},{"assignment":"11","diaID":"did@`@f\\bbbQ[hBB@A~dHBIU@","coupling":1.564,"multiplicity":"d"}]},{"atomIDs":["9"],"diaIDs":["did@`@f\\bbbQ[hBB@A~dHBIU@"],"integral":1,"delta":6.962,"atomLabel":"H","j":[{"assignment":"8","diaID":"did@`@fTfYea`H`@GzP`HeT","coupling":1.564,"multiplicity":"d"},{"assignment":"10","diaID":"did@`@fTfUff`@h@GzP`HeT","coupling":7.869,"multiplicity":"d"}]},{"atomIDs":["10"],"diaIDs":["did@`@fTfUff`@h@GzP`HeT"],"integral":1,"delta":7.173,"atomLabel":"H","j":[{"assignment":"8","diaID":"did@`@fTfYea`H`@GzP`HeT","coupling":0.519,"multiplicity":"d"},{"assignment":"9","diaID":"did@`@f\\bbbQ[hBB@A~dHBIU@","coupling":7.869,"multiplicity":"d"},{"assignment":"11","diaID":"did@`@f\\bbbQ[hBB@A~dHBIU@","coupling":7.869,"multiplicity":"d"}]},{"atomIDs":["11"],"diaIDs":["did@`@f\\bbbQ[hBB@A~dHBIU@"],"integral":1,"delta":6.962,"atomLabel":"H","j":[{"assignment":"8","diaID":"did@`@fTfYea`H`@GzP`HeT","coupling":1.564,"multiplicity":"d"},{"assignment":"10","diaID":"did@`@fTfUff`@h@GzP`HeT","coupling":7.869,"multiplicity":"d"}]},{"atomIDs":["12"],"diaIDs":["did@`@fTeeYnh@H@GzP`HeT"],"integral":1,"delta":2.237,"atomLabel":"H","j":[]},{"atomIDs":["13"],"diaIDs":["did@`@fTeeYnh@H@GzP`HeT"],"integral":1,"delta":2.237,"atomLabel":"H","j":[]},{"atomIDs":["14"],"diaIDs":["did@`@fTeeYnh@H@GzP`HeT"],"integral":1,"delta":2.237,"atomLabel":"H","j":[]},{"atomIDs":["15"],"diaIDs":["did@`@fTeeYnh@H@GzP`HeT"],"integral":1,"delta":2.237,"atomLabel":"H","j":[]},{"atomIDs":["16"],"diaIDs":["did@`@fTeeYnh@H@GzP`HeT"],"integral":1,"delta":2.237,"atomLabel":"H","j":[]},{"atomIDs":["17"],"diaIDs":["did@`@fTeeYnh@H@GzP`HeT"],"integral":1,"delta":2.237,"atomLabel":"H","j":[]}]
