--- src/range/__tests__/ranges.json
@@ -1,69 +1,0 @@
-[
-  {
-    "from": 2.9207911403931033,
-    "to": 3.005342719673898,
-    "integral": 0.8934763471847565,
-    "signal": [
-      {
-        "nbAtoms": 0,
-        "diaID": [],
-        "multiplicity": "dd",
-        "kind": "",
-        "remark": "",
-        "j": [
-          {
-            "coupling": 2.7202333074737908,
-            "multiplicity": "d"
-          },
-          {
-            "coupling": 9.176268879054533,
-            "multiplicity": "d"
-          }
-        ],
-        "delta": 2.9630669300335004
-      }
-    ]
-  },
-  {
-    "from": 2.702279993746368,
-    "to": 2.747747291380942,
-    "integral": 0.9345124056168826,
-    "signal": [
-      {
-        "nbAtoms": 0,
-        "diaID": [],
-        "multiplicity": "t",
-        "kind": "",
-        "remark": "",
-        "j": [
-          {
-            "coupling": 4.531576643179051,
-            "multiplicity": "t"
-          }
-        ],
-        "delta": 2.725013642563655
-      }
-    ]
-  },
-  {
-    "from": 1.2768799268371258,
-    "to": 1.3121201215410911,
-    "integral": 3.231412397661983,
-    "signal": [
-      {
-        "nbAtoms": 0,
-        "diaID": [],
-        "multiplicity": "d",
-        "kind": "",
-        "remark": "",
-        "j": [
-          {
-            "coupling": 5.194226872123636,
-            "multiplicity": "d"
-          }
-        ],
-        "delta": 1.2945000241891085
-      }
-    ]
-  }
-]
